CHANGED: renaming methods and imports
